task : finish calculator
